fix(backend): change backend directory structure
